Moved old code to avoid confusion.
